backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.1
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.31.0
